project/requirement.txt
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
jinja2
python-multipart
httpx
